requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.3.0
cached-property==1.5.2
chardet==3.0.4
idna==2.10
idna-ssl==1.1.0
mailru-im-async-bot==0.1
msgpack-python==0.5.6
multidict==4.7.6
pid==3.0.4
Pillow==8.0.1
psutil==5.7.3
six==1.15.0
tarantool==0.6.6
typing-extensions==3.7.4.3
yarl==1.6.3
